requirements.txt
streamlit==1.31.1
groq==0.9.0
langchain==0.1.16
langchain-community==0.0.34
langchain-text-splitters==0.0.1
sentence-transformers==2.6.1
faiss-cpu==1.9.0.post1
pypdf==4.2.0
python-dotenv==1.0.1
numpy==1.26.4
httpx==0.27.0
reportlab==4.1.0
pandas==2.2.2
streamlit-option-menu==0.3.6